backend/requirements.txt
# Python Dependencies for Django REST, PostgreSQL, Hugging Face, FAISS, and Celery

# Core Django
Django>=4.2
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-environ>=0.10.0

# Database
psycopg2-binary>=2.9
SQLAlchemy>=2.0

# SIEM Integration & APIs
requests>=2.31.0
requests-auth>=6.0.0

# Task Queue & Scheduling
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# AI/ML Dependencies
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
torch>=2.0.0

# Utilities
python-dotenv>=1.0.0
Faker>=18.0.0  # For test data generation
python-dateutil>=2.8.0

# Authentication
djangorestframework-simplejwt>=5.2.0
PyJWT>=2.8.0

# Monitoring & Logging
python-json-logger>=2.0.0
sentry-sdk>=1.25.0

# Production Server
gunicorn>=20.1.0
whitenoise>=6.4.0

# Testing
pytest>=7.3.0
pytest-django>=4.5.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
